state request in entrypoint + better exit codes

+++ images/openssl/configurationOptionsFactoryWithCoverage/buildScript.sh
@@ -7,11 +7,25 @@ make --directory=/src/openssl/ clean
 make --directory=/src/openssl/ install_sw
 EXIT_CODE=$?
 
-cp /build/lib/*.so.* /lib/
+if [ $EXIT_CODE -gt 0 ]
+then
+    exit $EXIT_CODE
+fi
+
+if [ -d "/build/lib/" ]; then
+    cp /build/lib/*.so.* /lib/
+else
+    exit 1
+fi
+
+if [ ! -f "/src/openssl/apps/openssl" ]; then
+    exit 1
+fi
+
 # Remove all non coverage file (but keep the openssl file). Used to keep the final cotainer small.
 find /src/openssl/ \( -not -name "*.gcno" -a -not -name "openssl" \) -type f -delete
 
 mkdir /usr/info
 printf "Build Date: $(date) \nBuild Parameter: $@" > /usr/info/build_info.txt
-exit $EXIT_CODE
+exit 0
